[PP][OPS-11967] mirgate EDH endpoint to backend
