+++ conf/application.conf
@@ -118,13 +118,6 @@ microservice {
       host = localhost
       port = 10153
     }
-    edh {
-      protocol = http
-      host = localhost
-      port = 10151
-      bearerToken = "test-bearer-token"
-      environment = "stub"
-    }
 
     casemanagement {
       protocol = http
